Add support for read replicas
* dbschema - if there are env vars prefixed with `HEROKU_POSTGRESQL`,
assume these are databases that are attached to the server,
and parse and feed into sequelize options as replicas. Sequelize will then
use one of the read replicas when making queries

* Mark db.transaction invocations as read-only, where appropriate

This commit was written while paired with @xkjyeah

This uses code taken from datagovsg/beeline-server-archive#214

NOTE: Read-replicas may not always catch up to master, so thus,
any changes made to master may not always be reflected in the replicas.
This might become a problem in situations where the user may want to see
a recent change made, eg, after a ticket purchase, or after an operator issues a refund