`/passengers` - use a better, more relevant sql query
`/passengers` queried the trip, then joined against route to authorize the caller
and a handful of tables via the TripTicket view to get the passenger manifest.

This was expensive, since that view joined against transactionItems using a text
match operator. The join in turn was needed to retrieve the flag `transactions.committed`,
which as it turns out, we don’t really need because the only time that is `false`
is if tickets are anything but `valid`.

Replace that query with two separate queries - one to look up just enough of the trip
to perform the authorization, and a raw query to grab just enough information for
`/passengers`

Performance comparisons verified using Postgres’ `EXPLAIN ANALYZE`